feat: semconv attribute search
# Why
We repeatedly find ourselves wanting to reference elements of the
semantic conventions. This is not easily doable with the semantic
conventions shown on opentelemetry.io.

 - It is hard to find where the definition for a specific attribute
   lives.
 - It is hard to deep link to it.

# What
Establish a list of all attributes with deep link support to a
specific attribute.

You could argue that this should be submitted to the OpenTelemetry
docs – this is likely the better way to handle this long term!

For now, this is a quick experiment.