feat: deep link API (#287)
* feat: deep link API

# Why

To integrate with the OTel GPT, it would be necessary that OTelBin
exposes an API to create deep links.

# What

Add an API under `POST /deep-link` that takes a collector config
within the request body and that outputs a deep link within the
`Location` response header and within the body (`text/plain` body).

* Update route.tsx

---------

Co-authored-by: Marcel Birkner <marcel.birkner@dash0.com>